refactor(val_el3,pal_el3): Standardize file naming with val_el3_ and pal_el3_ prefix
    - Renames all files in val_el3/ to follow consistent val_el3_ prefix.
    - Renames all files in pal_el3/ to follow consistent pal_el3_ prefix
    - Improves discoverability and enforces uniform naming convention
