+++ pal_el3/include/pal_el3_print.h
@@ -1,5 +1,22 @@
-#ifndef PAL_DEBUG_H
-#define PAL_DEBUG_H
+/** @file
+ * Copyright (c) 2023,2025, Arm Limited or its affiliates. All rights reserved.
+ * SPDX-License-Identifier : Apache-2.0
+
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+
+#ifndef PAL_EL3_PRINT_H
+#define PAL_EL3_PRINT_H
 
 #if defined(__ASSEMBLER__)
 # define   U(_x)        (_x)
@@ -100,5 +117,5 @@ extern void tf_log_newline(const char log_fmt[2]);
 extern void tf_log_set_max_level(unsigned int log_level);
 
 #endif /* __ASSEMBLER__ */
-#endif /* PAL_DEBUG_H */
+#endif /* PAL_EL3_PRINT_H */
