+++ val_el3/include/val_el3_security.h
@@ -1,0 +1,34 @@
+/** @file
+ * Copyright (c) 2025, Arm Limited or its affiliates. All rights reserved.
+ * SPDX-License-Identifier : Apache-2.0
+
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+
+#ifndef VAL_EL3_SECURITY_H
+#define VAL_EL3_SECURITY_H
+
+#include <val_el3_helpers.h>
+#include <val_el3_pe.h>
+#include <pal_el3.h>
+
+/* Prototypes moved from ack_common.c */
+#ifndef __ASSEMBLER__
+void val_enable_ns_encryption(void);
+void val_disable_ns_encryption(void);
+void val_prog_legacy_tz(int enable);
+void val_pas_filter_active_mode(int enable);
+void val_security_state_change(uint64_t attr_nse_ns);
+#endif /* __ASSEMBLER__ */
+
+#endif /* VAL_EL3_SECURITY_H */
