+++ val_el3/include/val_el3_smmu.h
@@ -1,3 +1,4 @@
+/* Unified SMMU header (v3 content folded) */
 /** @file
  * Copyright (c) 2025, Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
@@ -14,20 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-#ifndef __SMMU_V3_H__
-#define __SMMU_V3_H__
 
-#include <val_el3/ack_include.h>
+#ifndef VAL_EL3_SMMU_H
+#define VAL_EL3_SMMU_H
 
-#define BITFIELD_DECL(type, name, msb, lsb) \
-    const uint32_t name##_SHIFT = lsb; \
-    const type name##_MASK = ((((type)0x1) << (msb - lsb + 1)) - 1);
+#include <val_el3_helpers.h>
+#include <val_el3_memory.h>
+#include <val_el3_pgt.h>
+#include <val_el3_mec.h>
 
-#define BITFIELD_GET(name, val) ((val >> name##_SHIFT) & name##_MASK)
-#define BITFIELD_SET(name, val) ((val & name##_MASK) << name##_SHIFT)
-
-#define INPLACE(regfield, val) \
-        (((val) + UL(0)) << (regfield##_SHIFT))
 
 #define CMDQ_OP_PREFETCH_CFG 0x1
 #define CMDQ_OP_CFGI_STE 0x3
@@ -39,14 +35,12 @@
 
 #define BYTES_PER_DWORD 8
 
-BITFIELD_DECL(uint32_t, IDR0_ST_LEVEL, 28, 27)
 #define SMMU_IDR0_OFFSET 0x0
 #define IDR0_ST_LEVEL_2LVL 1
 #define IDR0_CD2L (1 << 19)
 #define IDR0_HYP (1 << 9)
 #define IDR0_COHACC (1 << 4)
 
-BITFIELD_DECL(uint32_t, IDR0_TTF, 3, 2)
 #define IDR0_TTF_AARCH64 2
 #define IDR0_TTF_AARCH32_64 3
 #define IDR0_S1P (1 << 1)
@@ -56,9 +50,6 @@ BITFIELD_DECL(uint32_t, IDR0_TTF, 3, 2)
 #define IDR1_TABLES_PRESET (1 << 30)
 #define IDR1_QUEUES_PRESET (1 << 29)
 #define IDR1_REL (1 << 28)
-BITFIELD_DECL(uint32_t, IDR1_CMDQS, 25, 21)
-BITFIELD_DECL(uint32_t, IDR1_SSIDSIZE, 10, 6)
-BITFIELD_DECL(uint32_t, IDR1_SIDSIZE, 5, 0)
 
 #define SMMU_OAS_MAX_IDX   7
 #define DPT_L0SZ_MAX_IDX   10
@@ -66,7 +57,6 @@ BITFIELD_DECL(uint32_t, IDR1_SIDSIZE, 5, 0)
 #define DPT_PS_MAX_IDX     8
 
 #define SMMU_IDR5_OFFSET 0x14
-BITFIELD_DECL(uint32_t, IDR5_OAS, 2, 0)
 
 #define SMMU_CR0_OFFSET 0x20
 #define CR0_EVENTQEN (1 << 2)
@@ -151,10 +141,10 @@ BITFIELD_DECL(uint32_t, IDR5_OAS, 2, 0)
 #define DPT_SHIFT		UL(15)
 #define MEC_SHIFT		UL(16)
 
-#define SMMUV3_FEAT_R_ATS	UL(1 << 1)
-#define SMMUV3_FEAT_R_MSI	UL(1 << 2)
-#define SMMUV3_FEAT_R_PRI	UL(1 << 3)
-#define SMMUV3_FEAT_R_MEC	UL(1 << 4)
+#define VAL_EL3_SMMU_H_
+#define VAL_EL3_SMMU_H_
+#define VAL_EL3_SMMU_H_
+#define VAL_EL3_SMMU_H_
 
 /* R_AIDR features */
 #define ARCH_REV_SHIFT 		UL(0)
@@ -169,12 +159,6 @@ BITFIELD_DECL(uint32_t, IDR5_OAS, 2, 0)
 #define EXTRACT(regfield, reg) \
         (((reg) & MASK(regfield)) >> (regfield##_SHIFT))
 
-BITFIELD_DECL(uint32_t, CR1_TABLE_SH, 11, 10)
-BITFIELD_DECL(uint32_t, CR1_TABLE_OC, 9, 8)
-BITFIELD_DECL(uint32_t, CR1_TABLE_IC, 7, 6)
-BITFIELD_DECL(uint32_t, CR1_QUEUE_SH, 5, 4)
-BITFIELD_DECL(uint32_t, CR1_QUEUE_OC, 3, 2)
-BITFIELD_DECL(uint32_t, CR1_QUEUE_IC, 1, 0)
 #define CR1_CACHE_NC 0
 #define CR1_CACHE_WB 1
 #define CR1_CACHE_WT 2
@@ -184,14 +168,10 @@ BITFIELD_DECL(uint32_t, CR1_QUEUE_IC, 1, 0)
 
 #define SMMU_STRTAB_BASE_OFFSET 0x80
 #define STRTAB_BASE_RA (1UL << 62)
-BITFIELD_DECL(uint64_t, STRTAB_BASE_ADDR, 51, 6)
 
 #define SMMU_STRTAB_BASE_CFG_OFFSET 0x88
-BITFIELD_DECL(uint32_t, STRTAB_BASE_CFG_FMT, 17, 16)
 #define STRTAB_BASE_CFG_FMT_LINEAR 0
 #define STRTAB_BASE_CFG_FMT_2LVL 1
-BITFIELD_DECL(uint32_t, STRTAB_BASE_CFG_SPLIT, 10, 6)
-BITFIELD_DECL(uint32_t, STRTAB_BASE_CFG_LOG2SIZE, 5, 0)
 
 #define SMMU_SH_NSH 0
 #define SMMU_SH_OSH 2
@@ -200,54 +180,30 @@ BITFIELD_DECL(uint32_t, STRTAB_BASE_CFG_LOG2SIZE, 5, 0)
 #define SMMU_MEMATTR_OIWB 0xf
 
 #define QUEUE_BASE_RWA (1UL << 62)
-BITFIELD_DECL(uint64_t, QUEUE_BASE_ADDR, 51, 5)
-BITFIELD_DECL(uint64_t, QUEUE_BASE_LOG2SIZE, 4, 0)
 
 #define STRTAB_L1_SZ_SHIFT 20
 #define STRTAB_SPLIT 8
 
 #define STRTAB_L1_DESC_DWORDS 1
 #define STRTAB_L1_DESC_SIZE 8
-BITFIELD_DECL(uint64_t, STRTAB_L1_DESC_SPAN, 4, 0)
-BITFIELD_DECL(uint64_t, STRTAB_L1_DESC_L2PTR, 51, 6)
 
 #define STRTAB_STE_DWORDS 8
 #define STRTAB_STE_0_V (1UL << 0)
-BITFIELD_DECL(uint64_t, STRTAB_STE_0_CONFIG, 3, 1)
 #define STRTAB_STE_0_CONFIG_ABORT 0
 #define STRTAB_STE_0_CONFIG_BYPASS 4
 #define STRTAB_STE_0_CONFIG_S1_TRANS 5
 #define STRTAB_STE_0_CONFIG_S2_TRANS 6
 
-BITFIELD_DECL(uint64_t, STRTAB_STE_0_S1FMT, 5, 4)
 #define STRTAB_STE_0_S1FMT_LINEAR 0
 #define STRTAB_STE_0_S1FMT_64K_L2 2
-BITFIELD_DECL(uint64_t, STRTAB_STE_0_S1CONTEXTPTR, 51, 6)
-BITFIELD_DECL(uint64_t, STRTAB_STE_0_S1CDMAX, 63, 59)
 
 #define STRTAB_STE_1_S1DSS_SSID0   0x2
 #define STRTAB_STE_1_S1C_CACHE_NC 0UL
 #define STRTAB_STE_1_S1C_CACHE_WBRA 1UL
-BITFIELD_DECL(uint64_t, STRTAB_STE_1_S1DSS, 1, 0)
-BITFIELD_DECL(uint64_t, STRTAB_STE_1_S1CIR, 3, 2)
-BITFIELD_DECL(uint64_t, STRTAB_STE_1_S1COR, 5, 4)
-BITFIELD_DECL(uint64_t, STRTAB_STE_1_S1CSH, 7, 6)
-BITFIELD_DECL(uint64_t, STRTAB_STE_1_EATS, 29, 28)
-BITFIELD_DECL(uint64_t, STRTAB_STE_1_STRW, 31, 30)
 #define STRTAB_STE_1_STRW_NSEL1 0UL
 #define STRTAB_STE_1_STRW_EL2 2UL
-BITFIELD_DECL(uint64_t, STRTAB_STE_1_SHCFG, 45, 44)
 #define STRTAB_STE_1_SHCFG_INCOMING 1UL
 
-BITFIELD_DECL(uint64_t, STRTAB_STE_2_S2VMID, 15, 0)
-BITFIELD_DECL(uint64_t, STRTAB_STE_2_VTCR, 50, 32)
-BITFIELD_DECL(uint64_t, STRTAB_STE_2_VTCR_S2T0SZ, 5, 0)
-BITFIELD_DECL(uint64_t, STRTAB_STE_2_VTCR_S2SL0, 7, 6)
-BITFIELD_DECL(uint64_t, STRTAB_STE_2_VTCR_S2IR0, 9, 8)
-BITFIELD_DECL(uint64_t, STRTAB_STE_2_VTCR_S2OR0, 11, 10)
-BITFIELD_DECL(uint64_t, STRTAB_STE_2_VTCR_S2SH0, 13, 12)
-BITFIELD_DECL(uint64_t, STRTAB_STE_2_VTCR_S2TG, 15, 14)
-BITFIELD_DECL(uint64_t, STRTAB_STE_2_VTCR_S2PS, 18, 16)
 
 #define STRTAB_STE_2_S2AA64 (1UL << 51)
 #define STRTAB_STE_2_S2PTW (1UL << 54)
@@ -297,11 +253,8 @@ BITFIELD_DECL(uint64_t, STRTAB_STE_2_VTCR_S2PS, 18, 16)
 #define VTCR_SL0_4K_L3          INPLACE(VTCR_SL0, UL(3))
 #define VTCR_SL0_4K_LM1         VTCR_SL0_4K_L2
 
-BITFIELD_DECL(uint64_t, STRTAB_STE_3_S2TTB, 51, 4)
 
 #define QUEUE_DWORDS_PER_ENT 2
-BITFIELD_DECL(uint64_t, CMDQ_0_OP, 7, 0)
-BITFIELD_DECL(uint64_t, CMDQ_CFGI_1_RANGE, 4, 0)
 #define CMDQ_CFGI_1_ALL_STES 31
 #define CMD_SID_SHIFT 		 U(32)
 #define CMD_SID_MASK 		 0xFFFFFFFF
@@ -313,26 +266,18 @@ BITFIELD_DECL(uint64_t, CMDQ_CFGI_1_RANGE, 4, 0)
 
 #define CDTAB_L1_DESC_DWORDS	1
 #define CDTAB_L1_DESC_V			(1UL << 0)
-BITFIELD_DECL(uint64_t, CDTAB_L1_DESC_L2PTR, 51, 12)
 
 #define CDTAB_CD_DWORDS		8
-BITFIELD_DECL(uint64_t, CDTAB_CD_0_TCR_T0SZ, 5, 0)
-BITFIELD_DECL(uint64_t, CDTAB_CD_0_TCR_TG0, 7, 6)
-BITFIELD_DECL(uint64_t, CDTAB_CD_0_TCR_IRGN0, 9, 8)
-BITFIELD_DECL(uint64_t, CDTAB_CD_0_TCR_ORGN0, 11, 10)
-BITFIELD_DECL(uint64_t, CDTAB_CD_0_TCR_SH0, 13, 12)
 #define CDTAB_CD_0_TCR_EPD0	(1ULL << 14)
 #define CDTAB_CD_0_TCR_EPD1	(1ULL << 30)
 
 #define CDTAB_CD_0_ENDI		(1UL << 15)
 #define CDTAB_CD_0_V		(1UL << 31)
 
-BITFIELD_DECL(uint64_t, CDTAB_CD_0_TCR_IPS, 34, 32)
 #define CDTAB_CD_0_AA64		(1UL << 41)
 #define CDTAB_CD_0_R		(1UL << 45)
 #define CDTAB_CD_0_A		(1UL << 46)
 #define CDTAB_CD_0_ASET		(1UL << 47)
-BITFIELD_DECL(uint64_t, CDTAB_CD_0_ASID, 63, 48)
 
 
 #define SMMU_STE_SIZE         64
@@ -342,16 +287,8 @@ BITFIELD_DECL(uint64_t, CDTAB_CD_0_ASID, 63, 48)
 #define STE_MECID_SHIFT       (STE_MECID_BIT % 64)
 #define STE_MECID_MASK        (0xFFFFULL << STE_MECID_SHIFT)
 
-BITFIELD_DECL(uint64_t, CDTAB_CD_1_TTB0, 51, 4)
 
-uint32_t smmu_oas[SMMU_OAS_MAX_IDX] = {32, 36, 40, 42, 44, 48, 52};
-uint32_t dpt_l0dptsz[DPT_L0SZ_MAX_IDX] = {30, 0, 0, 0, 34, 0, 36, 0, 0, 39};
-uint32_t l0dptsz_value[DPT_L0SZ_MAX_IDX] = {1, 0, 0, 0, 16, 0, 64, 0, 0, 512};
-uint32_t dpt_dptgs[DPT_GS_MAX_IDX] = {0x1000, 0x10000, 0x4000, 0};
-uint32_t dptgs_value[DPT_GS_MAX_IDX] = {4/1000000, 64/1000000, 16/1000000, 0};
-uint32_t dpt_dptps[DPT_PS_MAX_IDX] = {32, 36, 40, 42, 44, 58, 52, 0};
-uint32_t dptps_value[DPT_PS_MAX_IDX] = {4, 64, 1000, 4000, 16000, 256000, 4000000};
-
+#ifndef __ASSEMBLER__
 typedef struct {
     uint32_t prod;
     uint32_t cons;
@@ -458,7 +395,28 @@ typedef struct {
     uint32_t ssid_bits;
 } smmu_master_t;
 
+typedef struct {
+    uint32_t smmu_index;
+    uint32_t streamid;
+    uint32_t substreamid;
+    uint32_t ssid_bits;
+    uint32_t stage2;
+    uint32_t bypass;
+} smmu_master_attributes_t;
+
 uint32_t val_smmu_dpt_init(smmu_dev_t *smmu);
 int32_t smmu_reg_write_sync(smmu_dev_t *smmu, uint32_t val, uint32_t reg_off, uint32_t ack_off);
 void smmu_dpti_all(smmu_dev_t *smmu);
-#endif /*__SMMU_V3_H__ */
+void val_smmu_access_disable(uint64_t smmu_base);
+void val_smmu_access_enable(uint64_t smmu_base);
+void val_smmu_root_config_service(uint64_t arg0, uint64_t arg1, uint64_t arg2);
+void val_smmu_init_el3(uint32_t num_smmu, uint64_t smmu_base_arr[]);
+uint32_t val_smmu_rlm_map(smmu_master_attributes_t master_attr, pgt_descriptor_t pgt_desc);
+uint32_t val_dpt_add_entry(uint64_t translated_addr, uint64_t smmu_info);
+void val_dpt_invalidate_all(uint64_t smmu_index);
+uint32_t val_smmu_set_rlm_ste_mecid(smmu_master_attributes_t master_attr, uint32_t mecid);
+bool val_smmu_supports_mec(uint64_t smmu_base);
+uint32_t val_smmu_get_mecidw(uint64_t smmu_base);
+
+#endif /* __ASSEMBLER__ */
+#endif /* VAL_EL3_SMMU_H */
