+++ val_el3/src/val_el3_memory.c
@@ -1,0 +1,264 @@
+/** @file
+ * Copyright (c) 2025, Arm Limited or its affiliates. All rights reserved.
+ * SPDX-License-Identifier : Apache-2.0
+
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+
+#include <val_el3_debug.h>
+#include <val_el3_memory.h>
+#include <val_el3_mec.h>
+#include <val_el3_pe.h>
+
+struct_sh_data *shared_data = (struct_sh_data *) PLAT_SHARED_ADDRESS;
+
+static MemoryPool mem_pool = {
+    .base = (uint8_t *)PLAT_FREE_MEM_SMMU, // Hardcoded address
+    .size = PLAT_MEMORY_POOL_SIZE,
+    .free_list = NULL,
+};
+
+/**
+ * @brief  Returns the sligned address with the given size
+ *
+ * @param  size        Size in bytes to align
+ * @param  alignment   alignment required
+ * @return address aligned to the specified alignment till the 'size'
+ */
+static size_t align_size(size_t size, size_t alignment)
+{
+    return (size + (alignment - 1)) & ~(alignment - 1);
+}
+
+void map_shared_mem(void)
+{
+        INFO(" Function not implemented\n");
+}
+
+void val_data_cache_ops_by_va_el3(uint64_t VA, uint32_t type)
+{
+
+  switch (type)
+  {
+    case CLEAN_AND_INVALIDATE:
+      cln_and_invldt_cache((uint64_t *)VA);
+      break;
+    case CLEAN:
+      clean_cache((uint64_t *)VA);
+      break;
+    case INVALIDATE:
+      invalidate_cache((uint64_t *)VA);
+      break;
+    default:
+      shared_data->status_code = 1;
+      shared_data->error_code = 0;
+      const char *msg = "EL3: Invalid cache operation";
+      ERROR("\n %s", msg);
+      int i = 0; while (msg[i] && i < sizeof(shared_data->error_msg) - 1) {
+          shared_data->error_msg[i] = msg[i]; i++;
+      }
+      shared_data->error_msg[i] = '\0';
+      break;
+  }
+}
+
+void val_memory_set_el3(void *address, uint32_t size, uint8_t value)
+{
+  memset(address, value, size);
+}
+
+void val_memory_free_el3(void *ptr)
+{
+    uint32_t mecid = 0;
+
+    if (!ptr) return;
+
+   /* If MEC is enabled, the memory pool structures need to be accessed with
+       VAL_GMECID */
+    if (val_is_mec_enabled())
+    {
+        mecid = read_mecid_rl_a_el3();
+        val_write_mecid(VAL_GMECID);
+    }
+
+
+    BlockHeader *block = (BlockHeader *)((uint8_t *)ptr - sizeof(BlockHeader));
+    block->is_free = 1;
+
+    // Coalesce adjacent free blocks
+    BlockHeader *current = mem_pool.free_list;
+    while (current) {
+        if (current->is_free && current->next && current->next->is_free) {
+            current->size += current->next->size + sizeof(BlockHeader);
+            current->next = current->next->next;
+        }
+        current = current->next;
+    }
+
+    /* Restore MECID */
+    if (val_is_mec_enabled())
+        val_write_mecid(mecid);
+}
+
+void memory_pool_init(void)
+{
+    mem_pool.free_list = (BlockHeader *)mem_pool.base;
+    mem_pool.free_list->size = mem_pool.size - sizeof(BlockHeader);
+    mem_pool.free_list->is_free = 1;
+    mem_pool.free_list->next = NULL;
+}
+
+void split_block(BlockHeader *block, size_t size)
+{
+    BlockHeader *new_block = (BlockHeader *)((uint8_t *)block + sizeof(BlockHeader) + size);
+    new_block->size = block->size - size - sizeof(BlockHeader);
+    new_block->is_free = 1;
+    new_block->next = block->next;
+
+    block->size = size;
+    block->next = new_block;
+}
+
+/**
+  @brief   This function helps to read or write the address in EL3
+           1. Caller       -  Test Suite
+           2. Prerequisite -  Address needs to be mapped without any faults expected
+  @param   address - Address that needs to be read on or written on
+  @param   data    - The data which is written on the address
+  @return  None
+**/
+void access_mut(void)
+{
+  uint8_t type, num = shared_data->num_access;
+  uint64_t data;
+
+  for (int acc_cnt = 0; acc_cnt < num; ++acc_cnt)
+  {
+
+    type = shared_data->shared_data_access[acc_cnt].access_type;
+    switch (type)
+    {
+        case READ_DATA:
+          data = *(volatile uint32_t *) shared_data->shared_data_access[acc_cnt].addr;
+          VERBOSE("The data returned from the address, 0x%lx is 0x%x\n",
+               shared_data->shared_data_access[acc_cnt].addr, (uint32_t)data);
+          shared_data->shared_data_access[acc_cnt].data = data;
+          break;
+        case WRITE_DATA:
+          data = shared_data->shared_data_access[acc_cnt].data;
+          *(volatile uint32_t *)shared_data->shared_data_access[acc_cnt].addr = data;
+          VERBOSE("Data stored in VA, 0x%lx is 0x%x\n",
+                shared_data->shared_data_access[acc_cnt].addr,
+                *(uint32_t *)shared_data->shared_data_access[acc_cnt].addr);
+          break;
+        default:
+          ERROR("INVALID TYPE OF ACCESS");
+          break;
+    }
+  }
+}
+
+/**
+ * @brief  Allocates requested buffer size in bytes with zeros in a contiguous memory
+ *         and returns the base address of the range.
+ *
+ * @param  size         allocation size in bytes
+ * @param  num          Requested number of (buffer * size)
+ * @retval ptr          pointer to allocated memory
+ */
+void *val_memory_calloc_el3(size_t num, size_t size, size_t alignment)
+{
+    size_t total_size = num * size;
+    void *ptr = val_memory_alloc_el3(total_size, alignment);
+    if (ptr) {
+        memset(ptr, 0, total_size);
+    }
+    return ptr;
+}
+
+/**
+ * @brief  Returns the physical address of the requested Virtual address
+ *
+ * @param  Va  Virtual address
+ * @return Va  Returns the VA because of the 1:1 memory mapping
+ */
+void *val_memory_virt_to_phys_el3(void *va)
+{
+  return va;
+}
+
+/**
+ * @brief  Returns the virtual address of the requested physical address
+ *
+ * @param  pa  Physical address
+ * @return pa  Returns the PA because of the 1:1 memory mapping
+ */
+void *val_memory_phys_to_virt(uint64_t pa)
+{
+  return (void *)pa;
+}
+
+/**
+ * @brief  Allocates requested buffer size in bytes in a contiguous memory
+ *         and returns the base address of the range.
+ *
+ * @param  Size         allocation size in bytes
+ * @param  alignment    Required alignment for the buffer
+ * @retval if SUCCESS   pointer to allocated memory
+ * @retval if FAILURE   NULL
+ */
+void *val_memory_alloc_el3(size_t size, size_t alignment)
+{
+    uint32_t mecid = 0;
+
+    if (!mem_pool.free_list)
+    {
+        memory_pool_init(); // Initialize pool on first call
+    }
+
+    /* If MEC is enabled, the memory pool structures need to be accessed with
+       VAL_GMECID */
+    if (val_is_mec_enabled())
+    {
+        mecid = read_mecid_rl_a_el3();
+        val_write_mecid(VAL_GMECID);
+    }
+
+    size = align_size(size, alignment); // Align the requested size
+    BlockHeader *current = mem_pool.free_list;
+
+    while (current) {
+        // Align the starting address of the block
+        uintptr_t block_start = (uintptr_t)current + sizeof(BlockHeader);
+        uintptr_t aligned_start = align_size(block_start, alignment);
+        size_t alignment_padding = aligned_start - block_start;
+
+        if (current->is_free && current->size >= size + alignment_padding) {
+            if (current->size > size + alignment_padding + sizeof(BlockHeader)) {
+                split_block(current, size + alignment_padding);
+            }
+            current->is_free = 0;
+            /* Restore MECID */
+            if (val_is_mec_enabled())
+                val_write_mecid(mecid);
+            return (void *)aligned_start;
+        }
+        current = current->next;
+    }
+
+    /* Restore MECID */
+    if (val_is_mec_enabled())
+        val_write_mecid(mecid);
+
+    return NULL;
+}
